don't require a top level parent when preparing the patch set
This may be fragile. It works because we never actually need the top
level parent until we actually need to apply the patch set. When
prepping, either the top level parent is already referenced by other
nodes, or we have to create all the nodes and we can hanndle that in the
apply function.